refactor: added new source directory structure and makefile update
