--- Pipfile
@@ -1,25 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-autopep8 = "*"
-yapf = "*"
-
-[packages]
-gensim = "==3.8.1"
-pandas = "==0.20.1"
-scipy = "==1.3.0"
-numpy = "==1.17.0"
-seaborn = "==0.7.1"
-matplotlib = "==2.0.2"
-fuzzywuzzy = "==0.16.0"
-nltk = "==3.2.3"
-scikit-learn = "==0.20.0"
-ipykernel = "*"
-tqdm = "*"
-ipywidgets = "*"
-
-[requires]
-python_version = "3.6"
